Fix vendored deps for windows
